Fix SVG files.
They were previously using "currentColor" to fill all the black
sections, but currentColor only makes sense if defined by a previous
"color" property (which didn't exist). It seems that inkscape and
previous versions of librsvg just assume a default value of #000000 for for
"currentColor", but librsvg 2.40.2 in jessie draws it as transparent.

Explicitly specifying the color fixes this issue.